successful register controller test
