+++ src/main/resources/templates/success.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+    xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+    layout:decorate="~{fragments/layout}"
+    th:with="title=#{pages.login.name}"
+    th:lang="${lang}">
+<head>
+    <meta charset="UTF-8">
+</head>
+<body>
+    <main layout:fragment="content">
+        <p th:text="#{loginForm.text.pleaselogoutgoodsir}">Please visit</p>
+        <a href="/logout">Log out</a>
+    </main>
+</body>
+</html>
