--- src/test/java/se/kth/iv1201/group4/recruitment/application/PersonServiceTest.java
@@ -1,5 +1,0 @@
-package se.kth.iv1201.group4.application;
-
-public class PersonServiceTest {
-    
-}
